requirements_clean.txt
fastapi
uvicorn
pymongo
python-multipart
pypdf
langchain
langchain-community
langchain-core
langchain-groq
pinecone
langchain-google-genai
python-dotenv
passlib[bcrypt]
streamlit
requests
PyPDF2
tqdm
